refactor: move auth cookie interceptor to server
