+++ packages/client/package.json
@@ -17,7 +17,7 @@
     "test:watch": "jest ./ --watch"
   },
   "dependencies": {
-    "axios": "1.2.2",
+    "axios": "1.3.2",
     "classnames": "2.3.2",
     "dotenv": "16.0.2",
     "react": "18.2.0",
