+++ packages/server/middlewares/index.ts
@@ -2,3 +2,4 @@ export { jsonBodyParserMiddleware } from './bodyParser';
 export { checkAuthMiddleware } from './checkAuth';
 export { corsMiddleware } from './cors';
 export { proxyMiddleware } from './proxy';
+export { asyncLocalStorage, requestDataSaverMiddleware } from './requestDataSaver';
