+++ packages/server/package.json
@@ -11,6 +11,7 @@
     "test": "jest ."
   },
   "dependencies": {
+    "axios": "1.3.2",
     "body-parser": "1.20.1",
     "client": "0.0.0",
     "cors": "2.8.5",
